requirements/base.txt
asyncpg==0.21.0
newrelic==5.20.1.150
pylint==2.6.0
sanic==20.6.3
tortoise-orm==0.16.15
git+ssh://git@github.com/1mgOfficial/vessel.git@1.1.1